Make second argument of scheme_os_getcwd a size_t
By being an unsigned type, we ensure that when 0 is passed we know
that there's no smaller number. This aids the static analyzer.

+++ racket/src/racket/src/file.c
@@ -884,7 +884,7 @@ static Scheme_Object *string_to_path_element(int argc, Scheme_Object **argv)
 /*                                                                    */
 /**********************************************************************/
 
-char *scheme_os_getcwd(char *buf, int buflen, int *actlen, int noexn)
+char *scheme_os_getcwd(char *buf, size_t buflen, int *actlen, int noexn)
 {
   char *s;
   int slen;
